Remove Config class
