+++ src/common.ts
@@ -1,9 +1,10 @@
-import {Config} from './config';
+import {defaultStyles} from './config';
 import state from './state';
+import {assign} from "./polyfills";
 
 export function getStringWidth(text, styles) {
     let fontSize = styles.fontSize / state().scaleFactor;
-    Config.applyStyles(styles);
+    applyStyles(styles);
     text = Array.isArray(text) ? text : [text];
     let maxWidth = 0;
     text.forEach(function(line) {
@@ -48,7 +49,7 @@ export function ellipsize(text, width, styles, ellipsizeStr = '...') {
 export function addTableBorder() {
     let table = state().table;
     let styles = {lineWidth: table.settings.tableLineWidth, lineColor: table.settings.tableLineColor};
-    Config.applyStyles(styles);
+    applyStyles(styles);
     let fs = getFillStyle(styles);
     if (fs) {
         table.doc.rect(table.pageStartX, table.pageStartY, table.width, table.cursor.y - table.pageStartY, fs); 
@@ -67,4 +68,73 @@ export function getFillStyle(styles) {
     } else {
         return false;
     }
+}
+
+export function applyUserStyles() {
+    applyStyles(state().table.userStyles);
+}
+
+export function applyStyles(styles) {
+    let doc = state().doc;
+    let styleModifiers = {
+        fillColor: doc.setFillColor,
+        textColor: doc.setTextColor,
+        fontStyle: doc.setFontStyle,
+        lineColor: doc.setDrawColor,
+        lineWidth: doc.setLineWidth,
+        font: doc.setFont,
+        fontSize: doc.setFontSize
+    };
+    Object.keys(styleModifiers).forEach(function (name) {
+        let style = styles[name];
+        let modifier = styleModifiers[name];
+        if (typeof style !== 'undefined') {
+            if (Array.isArray(style)) {
+                modifier.apply(this, style);
+            } else {
+                modifier(style);
+            }
+        }
+    });
+}
+
+// This is messy, only keep array and number format the next major version
+export function marginOrPadding(value, defaultValue: number): any {
+    let newValue = {};
+    if (Array.isArray(value)) {
+        if (value.length >= 4) {
+            newValue = {'top': value[0], 'right': value[1], 'bottom': value[2], 'left': value[3]};
+        } else if (value.length === 3) {
+            newValue = {'top': value[0], 'right': value[1], 'bottom': value[2], 'left': value[1]};
+        } else if (value.length === 2) {
+            newValue = {'top': value[0], 'right': value[1], 'bottom': value[0], 'left': value[1]};
+        } else if (value.length === 1) {
+            value = value[0];
+        } else {
+            value = defaultValue;
+        }
+    } else if (typeof value === 'object') {
+        if (value['vertical']) {
+            value['top'] = value['vertical'];
+            value['bottom'] = value['vertical'];
+        } else if (value['horizontal']) {
+            value['right'] = value['horizontal'];
+            value['left'] = value['horizontal'];
+        }
+
+        for (let side of ['top', 'right', 'bottom', 'left']) {
+            newValue[side] = (value[side] || value[side] === 0) ? value[side] : defaultValue;
+        }
+    }
+
+    if (typeof value === 'number') {
+        newValue = {'top': value, 'right': value, 'bottom': value, 'left': value};
+    }
+
+    return newValue;
+}
+
+export function styles(styles) {
+    styles = Array.isArray(styles) ? styles : [styles];
+    return assign(defaultStyles(), ...styles);
 }
