+++ src/config.ts
@@ -8,7 +8,7 @@ import {assign} from './polyfills';
 /**
  * Styles for the themes (overriding the default styles)
  */
-export let getTheme = function(name) {
+export function getTheme(name) {
     let themes = {
         'striped': {
             table: {fillColor: 255, textColor: 80, fontStyle: 'normal'},
@@ -30,9 +30,9 @@ export let getTheme = function(name) {
         }
     };
     return themes[name];
-};
+}
 
-export function getDefaults() {
+export function defaultConfig() {
     return {
         // Styling
         theme: 'auto', // 'striped', 'grid' or 'plain'
@@ -76,7 +76,7 @@ export function getDefaults() {
 }
 
 // Base style for all themes
-function defaultStyles() {
+export function defaultStyles() {
     return {
         font: "helvetica", // helvetica, times, courier
         fontStyle: 'normal', // normal, bold, italic, bolditalic
@@ -92,80 +92,4 @@ function defaultStyles() {
         cellWidth: 'auto', // 'auto'|'wrap'|number
         minCellHeight: 0
     }
-}
-
-export class Config {
-    
-    static pageSize() {
-        return state().doc.internal.pageSize;
-    }
-    
-    static applyUserStyles() {
-        Config.applyStyles(state().table.userStyles);
-    }
-    
-    // This is messy, only keep array and number format the next major version
-    static marginOrPadding(value, defaultValue: number): any {
-        let newValue = {};
-        if (Array.isArray(value)) {
-            if (value.length >= 4) {
-                newValue = {'top': value[0], 'right': value[1], 'bottom': value[2], 'left': value[3]};
-            } else if (value.length === 3) {
-                newValue = {'top': value[0], 'right': value[1], 'bottom': value[2], 'left': value[1]};
-            } else if (value.length === 2) {
-                newValue = {'top': value[0], 'right': value[1], 'bottom': value[0], 'left': value[1]};
-            } else if (value.length === 1) {
-                value = value[0];
-            } else {
-                value = defaultValue;
-            }
-        } else if (typeof value === 'object') {
-            if (value['vertical']) {
-                value['top'] = value['vertical'];
-                value['bottom'] = value['vertical'];
-            } else if (value['horizontal']) {
-                value['right'] = value['horizontal'];
-                value['left'] = value['horizontal'];
-            }
-            
-            for (let side of ['top', 'right', 'bottom', 'left']) {
-                newValue[side] = (value[side] || value[side] === 0) ? value[side] : defaultValue;
-            }
-        }
-        
-        if (typeof value === 'number') {
-            newValue = {'top': value, 'right': value, 'bottom': value, 'left': value};
-        }
-        
-        return newValue;
-    }
-
-    static styles(styles) {
-        styles = Array.isArray(styles) ? styles : [styles];
-        return assign(defaultStyles(), ...styles);
-    }
-
-    static applyStyles(styles) {
-        let doc = state().doc;
-        let styleModifiers = {
-            fillColor: doc.setFillColor,
-            textColor: doc.setTextColor,
-            fontStyle: doc.setFontStyle,
-            lineColor: doc.setDrawColor,
-            lineWidth: doc.setLineWidth,
-            font: doc.setFont,
-            fontSize: doc.setFontSize
-        };
-        Object.keys(styleModifiers).forEach(function (name) {
-            let style = styles[name];
-            let modifier = styleModifiers[name];
-            if (typeof style !== 'undefined') {
-                if (Array.isArray(style)) {
-                    modifier.apply(this, style);
-                } else {
-                    modifier(style);
-                }
-            }
-        });
-    } 
 }
